packed gradient addition for histogram construction

+++ include/LightGBM/meta.h
@@ -40,7 +40,7 @@ typedef double score_t;
 typedef float score_t;
 #endif
 
-typedef int8_t int_score_t;
+typedef int32_t int_score_t;
 typedef int32_t int_hist_t;
 #define kIntGradBins 250
